basic: fix wrong reference / cast
num_in and rand_out are already `uint8_t*`. `(uint8_t*)&num_in` takes
the pointer of the pointer. It happens to return the same result for
the stack var, but it makes it hard to read and reason about. If the
rand_out var stops being a stack var, the behavior would change.

https://stackoverflow.com/questions/2528318/how-come-an-arrays-address-is-equal-to-its-value-in-c