Mark slot available when its key is deleted via C_DestroyObject.
Before this change, the object file would be deleted, but its slot would remain
marked as in-use. So a subsequence C_GenerateKeyPair would not find
any empty slots.

Also fix the code that finds an empty slot for a new key to return an error
if there are no empty slots. Before this change, it would pretend to succeed
and imply that slot 0 was available to write to.

Fixes #125