pkcs11: Fix deadlock in pkcs11_token_init
pkcs11_token_init() grabs the pContext->mutex at the beginning
of the function, locking the library. Later in the same function,
the code tries to grab the same lock, while already held. Mutexes
are not recursive in linux, resulting in a deadlock. Drop the
second attempt of grabbing the lock.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>