Fix symmetric_authenticate app.
The function symmetric_authenticate didn't work for the atsha204a. The
host_mac and device_mac did not match. This was due to the fact that
on the host side a key was derived whereas this was not done on the
device side.

The way I fixed it is to not do the key derivation on the host side,
this fixed the problem.