hal: zephyr: send word address when data buffer is empty
Some callers pass a word address, but empty data buffer, when issuing
i2c writes, such as when waking a device. Typically the HAL is passed a
data buffer that includes the word address as the first byte, but when
the data buffer is empty we opt to instead use the 1 byte word address
as the data to be written.

Signed-off-by: Daniel Mangum <georgedanielmangum@gmail.com>